+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## Rolling
+
+### Fixed
+
+- find_countries no longer look for countries codes by default
+
 ## [6.0.0] - 2026-04-07
 
 ### Added